4.8. Разделяй и властвуй
