+++ src/view/film-view.js
@@ -1,4 +1,4 @@
-import {createElement} from '../render';
+import AbstractView from './abstract-view';
 
 const createFilmCardTemplate = (film, comments) => {
   const {title, description, releaseDateYear, rating, runTime, genresOne, image, isWatchlist, isWatched, isFavorite} = film;
@@ -37,29 +37,18 @@ const createFilmCardTemplate = (film, comments) => {
 };
 
 
-export default class FilmCardView {
-  #element = null;
+export default class FilmCardView extends AbstractView {
   #films = null;
   #comments = null;
 
   constructor(films, comments) {
+    super();
     this.#films = films;
     this.#comments = comments;
-  }
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
   }
 
   get template() {
     return createFilmCardTemplate(this.#films, this.#comments);
   }
 
-  removeElement() {
-    this.#element = null;
-  }
 }
