+++ src/view/more-views.js
@@ -1,25 +1,11 @@
-import {createElement} from '../render';
+import AbstractView from './abstract-view.js';
 
 const createButtonMoreTemplate = () => (
   '<button class="films-list__show-more">Show more</button>'
 );
 
-export default class ButtonMoreView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
+export default class ButtonMoreView extends AbstractView {
   get template() {
     return createButtonMoreTemplate();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
