+++ src/view/statistics-view.js
@@ -1,28 +1,16 @@
-import {createElement} from '../render';
+import AbstractView from './abstract-view';
 
 const createFooterStatisticsTemplate = (films) => `<p>${films.length} movies inside</p>`;
 
-export default class FooterStatisticsView {
-  #element = null;
+export default class FooterStatisticsView extends AbstractView{
   #films = null;
 
   constructor(films) {
+    super();
     this.#films = films;
-  }
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
   }
 
   get template() {
     return createFooterStatisticsTemplate(this.#films);
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
